Chaos : Fix compilation errors
[FYI] jaco.vandyk
#rb trivial
#jira none

#ROBOMERGE-AUTHOR: cedric.caillaud
#ROBOMERGE-SOURCE: CL 18108172 in //UE4/Release-4.27-Chaos/... via CL 18108177 via CL 18108192
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v889-18060218)

[CL 18108203 by cedric caillaud in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/Experimental/Chaos/Public/Chaos/AABBTreeDirtyGridUtils.h
@@ -27,7 +27,7 @@ namespace Chaos
 		return (int64)(FMath::Floor(CellIndex + FloatingPointMaxCellIndexError));
 	}
 
-	FORCEINLINE_DEBUGGABLE int32 HashCell(int32 XCell, int32 YCell)
+	FORCEINLINE_DEBUGGABLE int32 HashCell(int64 XCell, int64 YCell)
 	{
 		return (int32)(InterleaveWithZeros((uint16)XCell) | (InterleaveWithZeros((uint16)YCell) << 1));
 	}
